Move most dialogs into `mapillary/gui/dialog`
