+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/actions/MapillaryUploadAction.java
@@ -13,7 +13,7 @@ import org.openstreetmap.josm.plugins.mapillary.MapillaryAbstractImage;
 import org.openstreetmap.josm.plugins.mapillary.MapillaryDataListener;
 import org.openstreetmap.josm.plugins.mapillary.MapillaryLayer;
 import org.openstreetmap.josm.plugins.mapillary.MapillaryPlugin;
-import org.openstreetmap.josm.plugins.mapillary.gui.MapillaryUploadDialog;
+import org.openstreetmap.josm.plugins.mapillary.gui.dialog.MapillaryUploadDialog;
 import org.openstreetmap.josm.plugins.mapillary.oauth.UploadUtils;
 import org.openstreetmap.josm.tools.I18n;
 import org.openstreetmap.josm.tools.ImageProvider;
