+++ src/main/java/org/openstreetmap/josm/plugins/mapillary/actions/MapillaryWalkAction.java
@@ -18,7 +18,7 @@ import org.openstreetmap.josm.plugins.mapillary.MapillaryDataListener;
 import org.openstreetmap.josm.plugins.mapillary.MapillaryLayer;
 import org.openstreetmap.josm.plugins.mapillary.MapillaryPlugin;
 import org.openstreetmap.josm.plugins.mapillary.gui.MapillaryMainDialog;
-import org.openstreetmap.josm.plugins.mapillary.gui.MapillaryWalkDialog;
+import org.openstreetmap.josm.plugins.mapillary.gui.dialog.MapillaryWalkDialog;
 import org.openstreetmap.josm.tools.ImageProvider;
 import org.openstreetmap.josm.tools.ImageProvider.ImageSizes;
